requirements.txt
arxiv==2.1.0
beautifulsoup4==4.12.3
gpt4all==2.4.0
lxml==5.2.1
numpy==1.26.4
openai==1.16.2
pandas==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
scenariogeneration==0.14.3
streamlit==1.33.0
